requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
python-dotenv
psycopg2-binary
greenlet
passlib[bcrypt]
bcrypt==4.0.1
jinja2
python-jose
pydantic[email]
python-multipart
faker